+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/cleanup/RobotUnusedVariableInspection.java
@@ -59,8 +59,11 @@ public class RobotUnusedVariableInspection extends LocalInspectionTool implement
         public void visitVariableDefinition(@NotNull RobotVariableDefinition o) {
             super.visitVariableDefinition(o);
             String name = o.getName();
-            if (name != null && !laterSeenVariableNames.contains(name) && definedVariableNames.add(name)) {
-                unusedVariables.put(name, o);
+            if (name != null) {
+                name = name.toLowerCase();
+                if (!laterSeenVariableNames.contains(name) && definedVariableNames.add(name)) {
+                    unusedVariables.put(name.toLowerCase(), o);
+                }
             }
         }
 
@@ -69,6 +72,7 @@ public class RobotUnusedVariableInspection extends LocalInspectionTool implement
             super.visitVariable(o);
             String variableName = o.getVariableName();
             if (variableName != null) {
+                variableName = variableName.toLowerCase();
                 ReadWriteAccessDetector detector = ReadWriteAccessDetector.findDetector(o);
                 if (detector != null && detector.getExpressionAccess(o) == Access.Read) {
                     if (unusedVariables.remove(variableName) == null) {